requirements.txt
numpy==2.2.4
Pillow==11.1.0
streamlit==1.42.0
tensorflow==2.19.0
